server/requirements.txt
RPi.GPIO
Flask
redis
pytz